==> 1_dask.ipynb <==
## One Model

rosshandler: https://pubmed.ncbi.nlm.nih.gov/37982461/ \
nowotschin: https://pubmed.ncbi.nlm.nih.gov/30959515/ \
see: https://scanpy.readthedocs.io/en/stable/tutorials/experimental/dask.html \
see: https://rapids-singlecell.readthedocs.io/en/latest/notebooks/demo_gpu-seuratv3-brain-1M.html

%%time
# importing our library 
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
#from scvi.models.utils import mde
import tempfile
import os
import scipy
from scipy.io import mmread
from scipy.io import mmwrite
import anndata as ad
import warnings 
import anndata as ad

import dask
import dask.array as da
#import dask.distributed as dd
import graphviz

from dask import delayed
import h5py


## Dask Setup

## setting graphics
dask.config.set({"visualization.engine": "graphviz"});
## Problem with memory spilling just solved itself
os.environ['MALLOC_TRIM_THRESHOLD_'] = '0'

#cluster = dd.LocalCluster(n_workers=24, memory_limit = 8*10**9)
#client = dd.Client(cluster)

#client

## Data loading & setting up anndata objects

chunk_size = [5000, -1]

%%time
## reading in the raw counts via raw/X
with h5py.File("/data/hadjantalab/atlas/extAtlas/embryo_complete.h5ad", "r") as f:
    adata_rosshandler = ad.AnnData(
        obs=ad.experimental.read_elem(f["obs"]),
        var=ad.experimental.read_elem(f["var"]),
        X = ad.experimental.read_elem(f["raw/X"])    
    )
adata_rosshandler.X = adata_rosshandler.X.astype(np.float32)
adata_rosshandler.X = da.from_array(adata_rosshandler.X)
adata_rosshandler.X = da.rechunk(adata_rosshandler.X, chunk_size)

## adata_rosshandler object
print(adata_rosshandler)
adata_rosshandler.X

# modulating some metadata
adata_rosshandler.var['mgi_symbol'] = adata_rosshandler.var['mgi_symbol'].astype('object')
adata_rosshandler.obs['batch'] = 'pijuan-sala'
adata_rosshandler.obs['batch'][adata_rosshandler.obs['embryo_version'] == 'Extension'] = 'rosshandler'
adata_rosshandler.var['id'] = adata_rosshandler.var_names
adata_rosshandler.var.index = adata_rosshandler.var['mgi_symbol']

%%time
with h5py.File("/data/hadjantalab/atlas/endoderm/sc_endoderm_all_cells.h5ad", "r") as f:
    adata_nowotschin = ad.AnnData(
        obs=ad.experimental.read_elem(f["obs"]),
        var=ad.experimental.read_elem(f["var"]),
        X = ad.experimental.read_elem(f["X"])    
    )

## reading in the raw counts via raw/X
adata_nowotschin.X = da.from_array(adata_nowotschin.X)
adata_nowotschin.X = da.rechunk(adata_nowotschin.X, chunk_size)

# adata_nowotschin object
print(adata_nowotschin)
adata_nowotschin.X

# modulating some metadata 
adata_nowotschin.obs['cell'] = adata_nowotschin.obs_names
adata_nowotschin.obs.index = adata_nowotschin.obs['cell']
adata_nowotschin.obs['celltype_nowotschin'] = adata_nowotschin.obs['CellType']
del adata_nowotschin.obs['CellType']
adata_nowotschin.obs['timepoint_nowotschin'] = adata_nowotschin.obs['Timepoint']
del adata_nowotschin.obs['Timepoint']
adata_nowotschin.obs['cluster_nowotschin'] = adata_nowotschin.obs['Cluster']
del adata_nowotschin.obs['Cluster']
adata_nowotschin.obs['batch'] = 'nowotschin'
adata_nowotschin.var['mgi_symbol'] = adata_nowotschin.var_names

## Concatenating with dask

## var.index unique? otherwise concat will not work
print('Nowotschin: ', len(adata_nowotschin.var.index) == len(set(adata_nowotschin.var.index)))
print('Rosshandler: ', len(adata_rosshandler.var.index) == len(set(adata_rosshandler.var.index)))


adata_rosshandler.var_names_make_unique()

%%time
import dask
from dask import delayed
import anndata as ad
import math

# Configure Dask to use multiple processors
dask.config.set(scheduler='threads', num_workers=10, memory_limit='230GB')
dask.config.set({'array.slicing.split_large_chunks': True})

# Define a function to perform delayed concatenation of all chunks at once
@delayed
def concatenate_all_chunks(chunks):
    return ad.concat(chunks, axis=0, join='outer', fill_value=0)

# Step 1: Determine appropriate chunk sizes
chunk_size_ross = int(math.ceil(adata_rosshandler.shape[0] / 5 / 1e4) * 1e4)
chunk_size_nowo = int(math.ceil(adata_nowotschin.shape[0] / 1e4) * 1e4)

# Step 2: Create chunks for both datasets
chunks_ross = [adata_rosshandler[i:i + chunk_size_ross] for i in range(0, len(adata_rosshandler), chunk_size_ross)]
chunks_nowo = [adata_nowotschin[i:i + chunk_size_nowo] for i in range(0, len(adata_nowotschin), chunk_size_nowo)]

# Step 3: Concatenate chunks from both datasets
# Concatenate both chunk lists at once (no need to pair them manually)
all_chunks = chunks_ross + chunks_nowo

# Step 4: Delayed final concatenation of all chunks
final_concatenated_data = concatenate_all_chunks(all_chunks)

# Compute the final concatenated anndata object
adata = dask.compute(final_concatenated_data)[0]

# Shape verification
expected_shape = (
    adata_rosshandler.shape[0] + adata_nowotschin.shape[0],
    len(set(adata_rosshandler.var_names).union(set(adata_nowotschin.var_names))),
)
print("Expected shape:", expected_shape)
print("Observed shape:", adata.shape)


final_concatenated_chunks.visualize()

final_concatenated_chunks.dask

adata

adata.X = adata.X.rechunk(chunk_size)
adata.X

adata.layers['dask'] = adata.X.copy()

adata.obs.index = adata.obs['cell']
del adata.obs['cell']

## Back to csr matrix

%%time
@dask.delayed
def get_chunk(c): 
    return c

chunks = adata.X.to_delayed().ravel()
results = [da.from_delayed(get_chunk(c), shape = chunk_size, dtype = np.float32) for c in chunks]
arr = da.concatenate(results, axis = 0, allow_unknown_chunksizes = True)

arr

#arr.visualize()

%%time
adata_x = arr.compute()

adata_x

adata.X = adata_x

## Saving data to disk

adata.obs.to_csv("/data/hadjantalab/lucas/atlas/data/dask/adata_obs.csv", index = 'cell')
adata.var.to_csv("/data/hadjantalab/lucas/atlas/data/dask/adata_var.csv")

%%time
mmwrite("/data/hadjantalab/lucas/atlas/data/dask/adata_X.mtx", a = adata_x)


def save_sparse_csr(filename, array):
    np.savez(filename, data=array.data, indices=array.indices,
             indptr=array.indptr, shape=array.shape)

%%time
save_sparse_csr("/data/hadjantalab/lucas/atlas/data/dask/adata_X.npz", adata_x)
