==> 2_adata_setup.ipynb <==
## One Model

rosshandler: https://pubmed.ncbi.nlm.nih.gov/37982461/ \
nowotschin: https://pubmed.ncbi.nlm.nih.gov/30959515/ \
https://www.nature.com/articles/s41592-021-01336-8 \
https://docs.scvi-tools.org/en/stable/tutorials/notebooks/scrna/harmonization.html \
https://rapids-singlecell.readthedocs.io/en/latest/notebooks/demo_gpu-seuratv3-brain-1M.html 

%%time
# importing our library 
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import scanpy as sc
#from scvi.models.utils import mde
import tempfile
import os
import scipy
from scipy.io import mmread
from scipy.io import mmwrite
import anndata as ad
import warnings 
import anndata as ad
import psutil
import h5py
import pickle
import sys

## Memory

## RAM info

# Get information about virtual memory usage
virtual_mem = psutil.virtual_memory()
# Access attributes like total, available, and used memory
total_mem = virtual_mem.total  # Total memory in bytes
available_mem = virtual_mem.available  # Available memory in bytes
used_mem = virtual_mem.used  # Used memory in bytes
# You can also calculate memory usage percentage
percent_used = used_mem / total_mem * 100

# printing it out
print(f"Total Memory: {total_mem / (1024**3)} GB")
print(f"Available Memory: {available_mem / (1024**3)} GB")
print(f"Used Memory: {used_mem / (1024**3)} GB")
print(f"Memory Usage Percentage: {percent_used:.2f}%")

## Setting Path

data_path = '/data/hadjantalab/lucas/atlas/data'

## LOADING METADATA

%%time
## loading rosshandler
## setting the dtype for pandas to speed up loading of csv

file_path = os.path.join(data_path, "dask/adata_obs.csv")
adata_obs = pd.read_csv(
    file_path, 
    index_col = 'cell', 
    dtype={
        'cell': object,
        'sample': 'category', 
           'embryo_version': 'category', 
           'stage' : 'category', 
           'somite_count' : 'category', 
           'S_score' : np.float32, 
           'louvain' : 'category', 
           'leiden' : 'category', 
           'celltype_PijuanSala2019' : 'category', 
           'celltype_extended_atlas' : 'category', 
           'celltype_nowotschin' : 'category',
           'timepoint_nowotschin': 'category', 
           'cluster_nowotschin': 'category', 
           'batch': 'category'
          }, 
    engine = 'c')

file_path = os.path.join(data_path, "dask/adata_var.csv")
adata_var = pd.read_csv(
    file_path, 
    index_col = 0,
    dtype={0:object},
    engine = 'c')


## NEW NAMING CONVENTIONS COLUMNS ETC

#### all timepoints in one column

# Create sets for each column
stage_set = set(adata_obs['stage'])
timepoint_set = set(adata_obs['timepoint_nowotschin'])

# Combine the sets using union
combined_set = stage_set.union(timepoint_set)

# Convert the combined set to a list
categories = pd.Series(list(combined_set)).dropna()

# setting this categorical dtype
cat_type = pd.CategoricalDtype(categories=categories, ordered=True)

# setting same category types
adata_obs['stage'] = adata_obs['stage'].astype(cat_type)
adata_obs['timepoint_nowotschin'] = adata_obs['timepoint_nowotschin'].astype(cat_type)

# transfer nowotschin timepoints to rosshandler timepoints
adata_obs['stage'][adata_obs['batch'] == 'nowotschin'] = adata_obs['timepoint_nowotschin'][adata_obs['batch'] == 'nowotschin']
## removing _ap and _gfp from stage
adata_obs['stage'][(adata_obs['stage'] == 'E8.75_ap') | (adata_obs['stage'] == 'E8.75_gfp')] = 'E8.75'
## removing the category type
adata_obs['stage'] = adata_obs['stage'].cat.remove_categories(['E8.75_ap', 'E8.75_gfp'])

## getting highly variable gene info from rosshandler and nowotschin

hvg_rosshandler = pd.read_csv(
    '/data/hadjantalab/atlas/extAtlas/metadata_genes.csv', 
    index_col = 'mgi_symbol', 
    dtype = {'highly_variable' : 'boolean'},
    usecols = ['highly_variable', 'mgi_symbol'],
    engine = 'c')
hvg_rosshandler = hvg_rosshandler[~hvg_rosshandler.index.duplicated(keep='first')]


hvg_nowotschin = pd.read_csv(
    '/data/hadjantalab/atlas/endoderm/HVGs.csv', 
    index_col = 'index',
    engine = 'c')
hvg_nowotschin['mgi_symbol'] = hvg_nowotschin.index
hvg_nowotschin.index = hvg_nowotschin['mgi_symbol']
del hvg_nowotschin['mgi_symbol']

print(f'Number of hvg in rosshandler', hvg_rosshandler['highly_variable'].sum())
print(f'Number of hvg in nowotschin', len(hvg_nowotschin))

hvg = list(hvg_rosshandler[hvg_rosshandler['highly_variable']].index)
for i in hvg_nowotschin.index:
    if i not in hvg:
        hvg.append(i)
print(f'Number of hvg outer join {len(hvg)}')

## adding highly variable gene info to our adata_var
adata_var['mgi_symbol'] = adata_var.index
adata_var.index = adata_var['mgi_symbol']
del adata_var['mgi_symbol']
adata_var['highly_variable'] = False
adata_var.index.isin(hvg)
adata_var['highly_variable'][adata_var.index.isin(hvg)] = True 

# TRYING TO GET SOME SAMPLE (SEQUENCING LANE) INFORMATION FOR THE NOWOTSCHIN DATASET

# Open the HDF5 file and read the data within the 'with' block
data_dict = {}  # Dictionary to store the data

with h5py.File('/data/hadjantalab/atlas/endoderm/manu_app_data/e8.5_metadata.h5', 'r') as f:
    # Get the first key in the file
    meta = list(f.keys())[0]
    # Access the group associated with the key
    meta_nowotschin = f[meta]

    if isinstance(meta_nowotschin, h5py.Group):
        print("This is a group. Keys in this group are:", list(meta_nowotschin.keys()))
        # Access and read all datasets within the group
        for key in meta_nowotschin.keys():
            dataset = meta_nowotschin[key]
            if isinstance(dataset, h5py.Dataset):
                data_dict[key] = dataset[...]
                print(f"Dataset {key} data: {data_dict[key]}")
    elif isinstance(meta_nowotschin, h5py.Dataset):
        # Read the dataset's data
        data_dict[meta] = meta_nowotschin[...]
        print("This is a dataset. Data is:", data_dict[meta])
    else:
        print("Unknown HDF5 object type.")

# Outside the 'with' block, you can now access 'data_dict' which contains the datasets' data
#print(data_dict)

### NO SEQUENCING LANE / SAMPLE INFO FOUND

## setting nowotschin dataset as celltype unknown for rosshandler annotation and vice versa
adata_obs['celltype_extended_atlas'] = adata_obs['celltype_extended_atlas'].cat.add_categories(['unknown'])
adata_obs['celltype_extended_atlas'][adata_obs['batch'] == 'nowotschin'] = 'unknown'
adata_obs['celltype_nowotschin'] = adata_obs['celltype_nowotschin'].cat.add_categories(['unknown'])
adata_obs['celltype_nowotschin'][adata_obs['batch'] != 'nowotschin'] = 'unknown'


## extracting sample info from the name of the nowotschin observations
nowo_sample_info = [x.rsplit('_', 1)[0] for x in adata_obs[adata_obs['batch'] == 'nowotschin'].index]
adata_obs['sample'] = adata_obs['sample'].cat.add_categories(set(nowo_sample_info))
## adding that information as sample info to adata_obs
adata_obs.loc[adata_obs['batch'] == 'nowotschin', 'sample'] = nowo_sample_info

### creating a compound column with the covariate batch, sample, and timepoint
adata_obs['covariate_composite'] = adata_obs['batch'].astype('string') + '_' + adata_obs['stage'].astype('string') + '_' + adata_obs['sample'].astype('string')

## deleting, renaming, and ordering

del adata_obs['embryo_version']
del adata_obs['timepoint_nowotschin']

adata_obs['celltype_rosshandler'] = adata_obs['celltype_extended_atlas'].copy()
del adata_obs['celltype_extended_atlas']

adata_obs['celltype_pijuansala'] = adata_obs['celltype_PijuanSala2019'].copy()
del adata_obs['celltype_PijuanSala2019']

adata_obs['rosshandler_somite_count'] = adata_obs['somite_count'].copy()
del adata_obs['somite_count']

adata_obs['rosshandler_anatomy'] = adata_obs['anatomy'].copy()
del adata_obs['anatomy']

adata_obs['rosshandler_S_score'] = adata_obs['S_score'].copy()
del adata_obs['S_score']

adata_obs['rosshandler_G2M_score'] = adata_obs['G2M_score'].copy()
del adata_obs['G2M_score']

adata_obs['rosshandler_phase'] = adata_obs['phase'].copy()
del adata_obs['phase']

adata_obs['rosshandler_louvain'] = adata_obs['louvain'].copy()
del adata_obs['louvain']

adata_obs['rosshandler_leiden'] = adata_obs['leiden'].copy()
del adata_obs['leiden']

adata_obs['nowotschin_cluster'] = adata_obs['cluster_nowotschin'].copy()
del adata_obs['cluster_nowotschin']


order = [ 'batch', 
         'stage', 
         'sample',
         'covariate_composite', 
         'celltype_nowotschin', 
         'celltype_rosshandler', 
         'celltype_pijuansala', 
         'nowotschin_cluster', 
         'rosshandler_somite_count', 
         'rosshandler_anatomy', 
         'rosshandler_phase',
         'rosshandler_S_score', 
         'rosshandler_G2M_score',
         'rosshandler_louvain', 
         'rosshandler_leiden']
adata_obs = adata_obs[order]

## RENAMING PROBLEMATIC ANNOTATION NAMES

adata_obs['celltype_rosshandler'] = adata_obs['celltype_rosshandler'].astype('object')
adata_obs.loc[adata_obs['celltype_rosshandler'] == 'NMPs/Mesoderm-biased', 'celltype_rosshandler'] = 'NMPs_Mesoderm-biased'
adata_obs.loc[adata_obs['celltype_rosshandler'] == 'Midbrain/Hindbrain boundary', 'celltype_rosshandler'] = 'Midbrain_Hindbrain boundary'

## LOADING COUNT TABLE

# function to read in our count matrix as csr_matrix
def load_sparse_csr(filename):
    loader = np.load(filename)
    return scipy.sparse.csr_matrix((loader['data'], loader['indices'], loader['indptr']), shape=loader['shape'])

%%time
file_path = os.path.join(data_path, "dask/adata_X.npz")

## reading in my csr_matrix
adata_x = load_sparse_csr(file_path)
print(f'X is a: {type(adata_x)} with the following shape: {adata_x.shape} and size: {(adata_x.data.nbytes + adata_x.indptr.nbytes + adata_x.indices.nbytes) /1e9} GB')

## ADATA SETUP 

# creating adata object
adata = ad.AnnData(X = adata_x, obs = adata_obs, var = adata_var)

# saving all the features (genes) in adata.raw
adata.raw = adata

# subsetting to our hvg
adata = adata[: ,adata.var['highly_variable']].copy()

## saving raw counts
adata.layers['raw_counts'] = adata.X.copy()

## getting some information on memory requirement 
print(f'X is a {type(adata.X)} with the following shape: {adata.X.shape} and size: {(adata.X.data.nbytes + adata.X.indptr.nbytes + adata.X.indices.nbytes) /1e9} GB')

# getting some more information on memory requirement

def get_total_size_of_anndata(adata):
    total_size = sys.getsizeof(adata)  # Base size of the AnnData object
    
    # Add sizes of the attributes
    total_size += sys.getsizeof(adata.X)         # Expression matrix
    total_size += sys.getsizeof(adata.obs)       # Observations
    total_size += sys.getsizeof(adata.var)       # Variables (genes)
    total_size += sys.getsizeof(adata.obsm)      # Multi-dimensional observations
    total_size += sys.getsizeof(adata.varm)      # Multi-dimensional variables
    total_size += sys.getsizeof(adata.uns)       # Unstructured annotations
    total_size += sys.getsizeof(adata.layers)    # Layers of data
    
    # Include the size of adata.raw, if it exists
    if adata.raw is not None:
        total_size += sys.getsizeof(adata.raw)
        total_size += sys.getsizeof(adata.raw.X)
        total_size += sys.getsizeof(adata.raw.var)
    
    return total_size / (1024 ** 3)  # Convert bytes to GB

# Example usage:
print(f"Total size of AnnData object (in RAM): {get_total_size_of_anndata(adata):.2f} GB")

## REMOVING 'MIXED_GASTRULATION' STAGE

mask = adata.obs['stage'] != 'Mixed gastrulation'
adata = adata[mask].copy()

## ADDING THE ANNOTATION


## creating new column
adata.obs['annotation'] = 'TO_BE_DETERMINED'
adata.obs.loc[adata.obs['batch'] == 'Sox17VEKO', 'annotation'] = 'To_be_determined'

## loading annotation table
data_path = '/data/hadjantalab/lucas/atlas/data'
file_path = os.path.join(data_path, 'annotation.csv')
annotation = pd.read_table(file_path, sep=',', index_col='cell')

## Converting annotation DataFrame to a dictionary with direct cell-to-annotation mapping
annotation = annotation.loc[:, 'annotation'].to_dict()

## Setting values in adata according to the annotation table
adata.obs.loc[adata.obs.index.isin(annotation.keys()), 'annotation'] = adata.obs.index.map(annotation).dropna()


adata.obs['annotation'] = adata.obs['annotation'].astype('object')
adata.obs.loc[adata.obs['annotation'] == 'NMPs/Mesoderm-biased', 'annotation'] = 'NMPs_Mesoderm-biased'
adata.obs.loc[adata.obs['annotation'] == 'Midbrain/Hindbrain boundary', 'annotation'] = 'Midbrain_Hindbrain boundary'

## CREATING 10E6 LOG2 NORMALIZATION

adata.X = adata.layers['raw_counts'].copy()

np.set_printoptions(threshold=np.inf)

subset = adata.X[10000:10050, 5000:5050].toarray()
print(subset)

np.set_printoptions(threshold=1000)  # or use np.set_printoptions() to reset to default


sc.pp.normalize_total(adata, target_sum=1e4)


np.set_printoptions(threshold=np.inf)

subset = adata.X[10000:10050, 5000:5050].toarray()
print(subset)

np.set_printoptions(threshold=1000)  # or use np.set_printoptions() to reset to default


sc.pp.log1p(adata, base = 2)


np.set_printoptions(threshold=np.inf)

subset = adata.X[10000:10050, 5000:5050].toarray()
print(subset)

np.set_printoptions(threshold=1000)  # or use np.set_printoptions() to reset to default


adata.layers['log_normalized'] = adata.X.copy()

## setting adata.X back to raw counts
adata.X = adata.layers['raw_counts'].copy()

## Save data 

#%%time
## saving everything seperately

## all metadata
#file_path = os.path.join(data_path, "adata_obs.csv")
#adata.obs.to_csv(file_path, index = 'cell')

#file_path = os.path.join(data_path, "adata_var.csv")
#adata.var.to_csv(file_path, index = 'mgi_symbol')

#def save_sparse_csr(filename, array):
#    np.savez(filename, data=array.data, indices=array.indices, indptr=array.indptr, shape=array.shape)

## saving sparse matrix seperately
#file_path = os.path.join(data_path, "adata_X.npz")
#save_sparse_csr(file_path, adata.X)

data_path

%%time
## saving the complete object, it ain't that big
file_path = os.path.join(data_path, "adata.h5ad")
adata.write_h5ad(file_path)

## LOAD DATA

data_path = '/data/hadjantalab/lucas/atlas/data'

%%time

file_path = os.path.join(data_path, "adata.h5ad")
adata = ad.read_h5ad(file_path)
