==> 3_covariates.ipynb <==
## for now only accounting for mitochondrial genes 

import anndata as ad
import os
import rpy2
import numpy as np
import pandas as pd
import rapids_singlecell as rsc
import sys
import scipy.sparse
import scanpy as sc



data_path = '/data/hadjantalab/lucas/atlas/data'

file_path = os.path.join(data_path, 'adata.h5ad')
adata = ad.read_h5ad(file_path)

adata

adata.obs['annotation_old'] = adata.obs['annotation'].copy()
del adata.obs['annotation']

# getting some more information on memory requirement

def get_total_size_of_anndata(adata):
    total_size = sys.getsizeof(adata)  # Base size of the AnnData object
    
    # Add sizes of the attributes
    total_size += sys.getsizeof(adata.X)         # Expression matrix
    total_size += sys.getsizeof(adata.obs)       # Observations
    total_size += sys.getsizeof(adata.var)       # Variables (genes)
    total_size += sys.getsizeof(adata.obsm)      # Multi-dimensional observations
    total_size += sys.getsizeof(adata.varm)      # Multi-dimensional variables
    total_size += sys.getsizeof(adata.uns)       # Unstructured annotations
    total_size += sys.getsizeof(adata.layers)    # Layers of data
    
    # Include the size of adata.raw, if it exists
    if adata.raw is not None:
        total_size += sys.getsizeof(adata.raw)
        total_size += sys.getsizeof(adata.raw.X)
        total_size += sys.getsizeof(adata.raw.var)
    
    return total_size / (1024 ** 3)  # Convert bytes to GB

# Example usage:
print(f"Total size of AnnData object (in RAM): {get_total_size_of_anndata(adata):.2f} GB")

## Cell cycle genes

## loading R to get list from satija lab

## setting R_HOME and R_USER
os.environ['R_HOME'] = '/home/schroel1/miniconda3/envs/gpu2/lib/R'
os.environ['R_USER'] = '/home/schroel1/miniconda3/envs/gpu2/lib/R'

# Import the rpy2 interface
import rpy2.robjects as ro
from rpy2.robjects import r


# Load the Seurat package in R
r('library(Seurat)')

## list from https://satijalab.org/seurat/archive/v3.0/cell_cycle_vignette.html 

# Define and execute the R code to access the gene sets
r('s_genes_satija <- cc.genes$s.genes')
r('g2m_genes_satija <- cc.genes$g2m.genes')

# Access the variables in Python
s_genes_satija = ro.globalenv['s_genes_satija']
g2m_genes_satija = ro.globalenv['g2m_genes_satija']

# convert them to Python objects
s_genes_satija = list(s_genes_satija)
g2m_genes_satija = list(g2m_genes_satija)

## combine the different cell cycle genes
cell_cycle_satija = s_genes_satija + g2m_genes_satija
# Convert both adata.var_names and cell_cycle_genes to lowercase for case-insensitive comparison
lowercase_mgi_symbols = adata.var_names.str.lower()
lowercase_cell_cycle_satija = [gene.lower() for gene in cell_cycle_satija]

# Perform the case-insensitive isin check
mask_cell_cycle_satija = lowercase_mgi_symbols.isin(lowercase_cell_cycle_satija)
filtered_cell_cycle_satija = list(adata.var_names[mask_cell_cycle_satija])
print(f'Genes in the list: {len(cell_cycle_satija)}, genes found in data: {len(filtered_cell_cycle_satija)}')


# list from https://yoseflab.github.io/scvi-tools-reproducibility/scvi_covariates/ 

cell_cycle_yosef = ['PCNA', 'dnk', 'RnrS', 'RnrL', 'Claspin', 'Mcm5', 'Caf1-180',
       'RPA2', 'HipHop', 'stg', 'Mcm6', 'dup', 'WRNexo', 'Mcm7', 'dpa',
       'CG10336', 'Mcm3', 'Mcm2', 'RpA-70', 'Chrac-14', 'CG13690', 'RPA3',
       'asf1', 'DNApol-alpha73', 'CycE', 'DNApol-alpha50',
       'Kmn1', 'Lam', 'Nph', 'msd5', 'msd1', 'ctp', 'Set', 'scra',
       'Chrac-16', 'ncd', 'Ote', 'pzg', 'HDAC1', 'nesd', 'tum', 'CG8173',
       'aurB', 'feo', 'pav', 'CG6767', 'sip2', 'Det', 'Cks30A', 'CycB',
       'B52']

# Convert both adata.var_names and cell_cycle_yosef to lowercase for case-insensitive comparison
lowercase_mgi_symbols = adata.var_names.str.lower()
lowercase_cell_cycle_yosef = [gene.lower() for gene in cell_cycle_yosef]

# Perform the case-insensitive isin check
filtered_cell_cycle_yosef = adata.var_names[lowercase_mgi_symbols.isin(lowercase_cell_cycle_yosef)]
## check the genes found
print(f'Genes in the list: {len(cell_cycle_yosef)}, genes found in data: {len(filtered_cell_cycle_yosef)}')
filtered_cell_cycle_yosef


## converting to a python list
filtered_cell_cycle_yosef = list(filtered_cell_cycle_yosef)


# Get the indices of all the genes in `filtered_cell_cycle_satija`
#gene_indices = adata.var.index.get_indexer(filtered_cell_cycle_satija)

# Extract all columns corresponding to these gene indices from the sparse matrix at once
#genes_expression = adata.layers['log_normalized'][:, gene_indices]

# Convert to dense matrix only if necessary (e.g., for adding to adata.obs)
# Instead of converting one column at a time, convert the entire block at once
#genes_expression_dense = genes_expression.toarray()

# Add each gene's expression to adata.obs, all at once
#for i, gene in enumerate(filtered_cell_cycle_satija):
#    adata.obs[gene] = genes_expression_dense[:, i]


#adata._inplace_subset_var(~mask_cell_cycle_satija)

## Mitochondrial genes

lowercase_mgi_symbols = adata.var_names.str.lower()
mask_mito = lowercase_mgi_symbols.str.startswith('mt-')
filtered_mito = list(adata.var_names[mask_mito])
filtered_mito

# Get the indices of all the genes in `filtered_cell_cycle_satija`
gene_indices = adata.var.index.get_indexer(filtered_mito)

# Extract all columns corresponding to these gene indices from the sparse matrix at once
genes_expression = adata.layers['log_normalized'][:, gene_indices]

# Convert to dense matrix only if necessary (e.g., for adding to adata.obs)
# Instead of converting one column at a time, convert the entire block at once
genes_expression_dense = genes_expression.toarray()

# Add each gene's expression to adata.obs, all at once
for i, gene in enumerate(filtered_mito):
    adata.obs[gene] = genes_expression_dense[:, i]


adata._inplace_subset_var(~mask_mito)

## Ribosomal genes

lowercase_mgi_symbols = adata.var_names.str.lower()
mask_ribo = lowercase_mgi_symbols.str.startswith('rps') | lowercase_mgi_symbols.str.startswith('rpl')
filtered_ribo = list(adata.var_names[mask_ribo])
filtered_ribo

## Pseudogenes

lowercase_mgi_symbols = adata.var_names.str.lower()
mask_pseudogenes = lowercase_mgi_symbols.str.match(r'^gm\d+$')
filtered_pseudogenes = list(adata.var_names[mask_pseudogenes])
filtered_pseudogenes

#adata._inplace_subset_var(~mask_pseudogenes)

## Sex related genes

## import r line magic
#%load_ext rpy2.ipython
#%R library(biomaRt)

#%R print(listAttributes(mart)$name)

## filter the list on the 
#%R mart <- useMart(biomart="ensembl", dataset="mmusculus_gene_ensembl")
#%R genes_sex_male <- getBM(attributes = c("mgi_symbol"), filters = "chromosome_name", values = "Y", mart = mart)
#%R genes_sex_male

#genes_sex_male = ro.globalenv['genes_sex_male']
#genes_sex_male = list(genes_sex_male[0])
#genes_sex_female = ['Xist', 'Tsix']
#genes_sex = genes_sex_male + genes_sex_female

#lowercase_mgi_symbols = adata.var_names.str.lower()
#mask_sex = lowercase_mgi_symbols.isin(str.lower(gene) for gene in genes_sex)
#filtered_sex = list(adata.var_names[mask_sex])
#filtered_sex

# Get the indices of all the genes in `filtered_cell_cycle_satija`
#gene_indices = adata.var.index.get_indexer(filtered_sex)

# Extract all columns corresponding to these gene indices from the sparse matrix at once
#genes_expression = adata.layers['log_normalized'][:, gene_indices]

# Convert to dense matrix only if necessary (e.g., for adding to adata.obs)
# Instead of converting one column at a time, convert the entire block at once
#genes_expression_dense = genes_expression.toarray()

# Add each gene's expression to adata.obs, all at once
#for i, gene in enumerate(filtered_sex):
#    adata.obs[gene] = genes_expression_dense[:, i]


#adata._inplace_subset_var(~mask_sex)

## Chemistry

## adding chemistry did not change a thing

#adata.obs['chemistry'] = 'to_be_determined'
#adata.obs.loc[adata.obs['batch'] == 'pijuan-sala', 'chemistry'] = 'v1'
#adata.obs.loc[adata.obs['batch'] == 'nowotschin', 'chemistry'] = 'v2'
#adata.obs.loc[adata.obs['batch'] == 'rosshandler', 'chemistry'] = 'v3'
#adata.obs.loc[adata.obs['batch'] == 'pulse_chase_labeled', 'chemistry'] = 'v3'
#adata.obs['chemistry'].value_counts()

## Cell cycle score

s_genes_satija = [gene.capitalize() for gene in s_genes_satija]
g2m_genes_satija = [gene.capitalize() for gene in g2m_genes_satija]

print(sum(adata.var_names.isin(s_genes_satija)))
print(sum(adata.var_names.isin(g2m_genes_satija)))

## USING RAW INSTEAD TO GET MORE INFORMATION
print(sum(adata.raw.var_names.isin(s_genes_satija)))
print(sum(adata.raw.var_names.isin(g2m_genes_satija)))

s_genes = adata.raw.var_names[adata.raw.var_names.isin(s_genes_satija)]
g2m_genes = adata.raw.var_names[adata.raw.var_names.isin(g2m_genes_satija)]


%%time
sc.tl.score_genes_cell_cycle(adata, s_genes = s_genes, g2m_genes = g2m_genes)

## Saving adata object

print(f"Total size of AnnData object (in RAM): {get_total_size_of_anndata(adata):.2f} GB")

data_path = '/data/hadjantalab/lucas/atlas/data'

%%time
## saving the complete object, it ain't that big
file_path = os.path.join(data_path, "adata.h5ad")
adata.write_h5ad(file_path)
