==> 8_reanno.ipynb <==
import scanpy as sc
import anndata as ad
import pandas as pd
import os
import numpy as np

## LOADING DATA

data_path = '/data/hadjantalab/lucas/atlas/data/reannotation'

file_path = os.path.join(data_path, 'reanno_endo_cohort55-775_nn-transfer.csv')
reanno_endo_cohort55_775_nn_transfer = pd.read_table(file_path, sep = ',', index_col = 'cell' )
reanno_endo_cohort55_775_nn_transfer = reanno_endo_cohort55_775_nn_transfer.dropna()

file_path = os.path.join(data_path, 'reanno_endo_cohort80-95_nn-transfer.csv')
reanno_endo_cohort80_95_nn_transfer = pd.read_table(file_path, sep = ',', index_col = 'cell' )
reanno_endo_cohort80_95_nn_transfer = reanno_endo_cohort80_95_nn_transfer.dropna()

file_path = os.path.join(data_path, 'reanno_meso_cohort55-775_nn-transfer.csv')
reanno_meso_cohort55_775_nn_transfer = pd.read_table(file_path, sep = ',', index_col = 'cell' )
reanno_meso_cohort55_775_nn_transfer = reanno_meso_cohort55_775_nn_transfer.dropna()

file_path = os.path.join(data_path, 'reanno_meso_cohort80-95_nn-transfer.csv')
reanno_meso_cohort80_95_nn_transfer = pd.read_table(file_path, sep = ',', index_col = 'cell' )
reanno_meso_cohort80_95_nn_transfer = reanno_meso_cohort80_95_nn_transfer.dropna()

nowotschin_blood = pd.DataFrame(adata.obs.loc[adata.obs['celltype_nowotschin'] == 'Blood', 'celltype_nowotschin'].copy())
nowotschin_blood['nn_transfer'] = nowotschin_blood['celltype_nowotschin'].copy()
del nowotschin_blood['celltype_nowotschin']
nowotschin_blood['nn_transfer'] = 'Erythroid'

data_path = '/data/hadjantalab/lucas/atlas/data'
file_path = os.path.join(data_path, 'adata.h5ad')
adata = ad.read_h5ad(file_path)

## ADDING TO ADATA

reanno = pd.concat([reanno_endo_cohort55_775_nn_transfer, reanno_endo_cohort80_95_nn_transfer, reanno_meso_cohort55_775_nn_transfer, reanno_meso_cohort80_95_nn_transfer, nowotschin_blood])

## CREATING NEW ANNOTATION COLUMN AND FILLING IT WITH VALUES

adata.obs['annotation'] = 'to_be_determined'
adata.obs['annotation'] = adata.obs['annotation'].astype('object')

# Use .loc to modify the 'annotation' column for rows where 'celltype_nowotschin' is not 'unknown'
adata.obs.loc[adata.obs['celltype_nowotschin'] != 'unknown', 'annotation'] = \
    adata.obs.loc[adata.obs['celltype_nowotschin'] != 'unknown', 'celltype_nowotschin']

# Use .loc to modify the 'annotation' column for rows where 'celltype_rosshandler' is not 'unknown'
adata.obs.loc[adata.obs['celltype_rosshandler'] != 'unknown', 'annotation'] = \
    adata.obs.loc[adata.obs['celltype_rosshandler'] != 'unknown', 'celltype_rosshandler']



with pd.option_context('display.max_rows', None):
    print(adata.obs['annotation'].value_counts(dropna=False).sort_index())


len(np.unique(adata.obs['annotation']))

## TRANSFERRING THE NEW LABELS 
adata.obs.loc[reanno.index, 'annotation'] = reanno['nn_transfer'].copy()


len(np.unique(adata.obs['annotation']))

with pd.option_context('display.max_rows', None):
    print(adata.obs['annotation'].value_counts(dropna=False).sort_index())

adata

## SAVE DATA 

data_path = '/data/hadjantalab/lucas/atlas/data'
file_path = os.path.join(data_path, 'adata.h5ad')
adata.write_h5ad(file_path)

data_path = '/data/hadjantalab/lucas/atlas/data/reannotation'
file_path = os.path.join(data_path, 'reannotation.csv')
reanno.to_csv(file_path, sep = ',', index = True)

anno = pd.DataFrame(adata.obs['annotation'].copy())
data_path = '/data/hadjantalab/lucas/atlas/data'
file_path = os.path.join(data_path, 'annotation.csv')
anno.to_csv(file_path, sep = ',', index = True)
